accountingextf-files-client/pyproject.toml
[tool.poetry]
name = "accountingextf-files-client"
version = "3.0"
description = "A client library for accessing accounting:extf-files"
authors = []
readme = "README.md"
packages = [
    {include = "accountingextf_files_client"},
]
include = ["CHANGELOG.md", "accountingextf_files_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.20.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
